requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0
webdriver-manager>=3.8.0

# Data processing and ML
scikit-learn>=1.1.0
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0
open_clip_torch>=2.20.0

# Vector database
chromadb>=0.4.0
faiss-cpu>=1.7.0

# API and web framework
flask>=2.2.0
openai>=1.0.0
ollama>=0.1.0

# Image processing
Pillow>=9.2.0
opencv-python>=4.6.0

# Data download
kagglehub>=0.2.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
